Add Docker CLI version check.

+++ script/dockerized.sh
@@ -28,13 +28,18 @@ PROJECT_DIR=$(cd ${0%/*}/..; pwd)
 BuildEnvironment=-$1; shift
 BuildEnvironment=${BuildEnvironment/-base}
 
-docker run -it --rm -h fishtank \
-    -e HOST_UID=$(id -u) -e HOST_GID=$(id -g) \
-    --env-file $PROJECT_DIR/.env-file \
-    --mount type=bind,source=$PROJECT_DIR,target=/project_dir \
-    --mount source=ccache_dir,target=/ccache_dir \
-    --mount source=home_dir,target=/home/urho3d \
-    --name dockerized$BuildEnvironment \
-    urho3d/dockerized$BuildEnvironment $@
+if [[ $(docker version -f {{.Client.Version}}) =~ ^([0-9]+)\.0*([0-9]+)\. ]] && (( ${BASH_REMATCH[1]} * 100 + ${BASH_REMATCH[2]} >= 1809 )); then
+    docker run -it --rm -h fishtank \
+        -e HOST_UID=$(id -u) -e HOST_GID=$(id -g) \
+        --env-file $PROJECT_DIR/.env-file \
+        --mount type=bind,source=$PROJECT_DIR,target=/project_dir \
+        --mount source=ccache_dir,target=/ccache_dir \
+        --mount source=home_dir,target=/home/urho3d \
+        --name dockerized$BuildEnvironment \
+        urho3d/dockerized$BuildEnvironment $@
+else
+    echo "Error: Docker CLI version is too old, minimum required version is 18.09.x"
+    exit 2;
+fi
 
 # vi: set ts=4 sw=4 expandtab:
